Made minor folder restructures to folders within app/features
