+++ projects/app/src/app/features/footers/components/footer-default/footer-default.component.scss
