+++ projects/app/src/app/features/footers/components/index.ts
@@ -1,0 +1,1 @@
+export * from './footer-default/footer-default.component';
