+++ projects/app/src/app/features/headers/components/index.ts
@@ -1,0 +1,1 @@
+export * from './header-default/header-default.component';
