+++ projects/app/src/app/features/headers/index.ts
@@ -1,1 +1,3 @@
+export * from './components';
+
 export * from './headers.module';
