+++ projects/app/src/app/features/layouts/components/index.ts
@@ -1,0 +1,1 @@
+export * from './layout-default/layout-default.component';
