+++ projects/app/src/app/features/layouts/components/layout-default/layout-default.component.scss
