+++ projects/app/src/app/features/layouts/index.ts
@@ -1,1 +1,3 @@
+export * from './components';
+
 export * from './layouts.module';
